Add dask-submit cli
